move common analysis to standard analysis module
