+++ qiskit_experiments/curve_analysis/standard_analysis/__init__.py
@@ -10,6 +10,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Calibration Analysis Classes"""
+"""Standard curve analysis library."""
 
-from .drag_analysis import DragCalAnalysis
+from .oscillation import OscillationAnalysis
+from .resonance import ResonanceAnalysis
