+++ qiskit_experiments/curve_analysis/standard_analysis/oscillation.py
