+++ qiskit_experiments/curve_analysis/standard_analysis/resonance.py
