+++ qiskit_experiments/library/calibration/__init__.py
@@ -52,7 +52,6 @@ Calibration analysis
     :toctree: ../stubs/
     :template: autosummary/analysis.rst
 
-    OscillationAnalysis
     DragCalAnalysis
     FineAmplitudeAnalysis
 
@@ -67,7 +66,6 @@ from .rabi import Rabi, EFRabi
 from .fine_amplitude import FineAmplitude, FineXAmplitude, FineSXAmplitude
 from .ramsey_xy import RamseyXY
 
-from .analysis.oscillation_analysis import OscillationAnalysis
 from .analysis.drag_analysis import DragCalAnalysis
 from .analysis.fine_amplitude_analysis import FineAmplitudeAnalysis
 from .analysis.remsey_xy_analysis import RamseyXYAnalysis
