+++ qiskit_experiments/library/calibration/rabi.py
@@ -22,8 +22,7 @@ from qiskit.providers import Backend
 import qiskit.pulse as pulse
 
 from qiskit_experiments.framework import BaseExperiment, Options
-from qiskit_experiments.curve_analysis import ParameterRepr
-from qiskit_experiments.library.calibration.analysis.oscillation_analysis import OscillationAnalysis
+from qiskit_experiments.curve_analysis import ParameterRepr, OscillationAnalysis
 from qiskit_experiments.exceptions import CalibrationError
