+++ qiskit_experiments/library/characterization/__init__.py
@@ -39,7 +39,6 @@ Analysis
 
     T1Analysis
     T2RamseyAnalysis
-    ResonanceAnalysis
     CrossResonanceHamiltonianAnalysis
 
 """
@@ -47,7 +46,6 @@ Analysis
 from .t1 import T1
 from .t1_analysis import T1Analysis
 from .qubit_spectroscopy import QubitSpectroscopy
-from .resonance_analysis import ResonanceAnalysis
 from .ef_spectroscopy import EFSpectroscopy
 from .t2ramsey import T2Ramsey
 from .t2ramsey_analysis import T2RamseyAnalysis
